Fixed an error where it wasn`t possible to get the authorization context when redirecting
When redirecting to a new page with the entire query string saved, an error was returned that redirect_uri was too long or missing. However, if you enter the same request into the address bar, i.e. do not redirect, but as if follow the link, then the authorization context is issued without problems.